rent history

+++ Carly/src/redux/api.js
@@ -10,6 +10,8 @@ import {
   topUpSuccess,
   likeCar,
   unlikeCar,
+  getRentHistoryCars,
+  loginAgain,
 } from './actions';
 
 const URL = 'https://wedcarly.azurewebsites.net';
@@ -60,6 +62,33 @@ export const login =
       throw error; // Rethrow the error for the calling function to handle
     }
   };
+export const logAgain = () => async (dispatch) => {
+  var userInfo = await AsyncStorage.getItem('userInfo');
+  userInfo = JSON.parse(userInfo);
+  console.log(userInfo);
+
+  try {
+    const response = await axios.post(`${URL}/auth/login`, {
+      username: 'abc',
+      password: 'abc',
+    });
+
+    console.log(response + "xdxd");
+
+    if (response.status === 200) {
+      const jwtToken = response.data.jwttoken;
+      dispatch(loginSuccess({ ...userInfo, jwtToken }));
+
+      await AsyncStorage.setItem('userInfo', JSON.stringify({ ...userInfo, jwtToken }));
+    } else {
+      throw new Error('Login failed');
+    }
+  } catch (error) {
+    console.error('Error during login:', error);
+    throw error; // Rethrow the error for the calling function to handle
+  }
+};
+
 export const register =
   ({ username, firstName, lastName, email, password }) =>
   async (dispatch) => {
@@ -190,7 +219,12 @@ export const getPayments = () => async (dispatch) => {
 };
 
 export const sendLikedCar = (id) => async (dispatch) => {
-  const jwtToken = useSelector((state) => state.userInfo.jwtToken);
+  const userInfo = await AsyncStorage.getItem('userInfo');
+  const { jwtToken } = JSON.parse(userInfo);
+
+  if (!jwtToken) {
+    throw new Error('JWT token not found. User must be logged in.');
+  }
   try {
     const response = await axios.post(`${URL}/users/favorites/${id}`, {
       headers: {
@@ -210,7 +244,12 @@ export const sendLikedCar = (id) => async (dispatch) => {
 };
 
 export const sendUnlikedCar = (id) => async (dispatch) => {
-  const jwtToken = useSelector((state) => state.userInfo.jwtToken);
+  const userInfo = await AsyncStorage.getItem('userInfo');
+  const { jwtToken } = JSON.parse(userInfo);
+
+  if (!jwtToken) {
+    throw new Error('JWT token not found. User must be logged in.');
+  }
   try {
     const response = await axios.delete(`${URL}/users/favorites/${id}`, {
       headers: {
@@ -230,14 +269,19 @@ export const sendUnlikedCar = (id) => async (dispatch) => {
 };
 
 export const fetchFavoriteCars = () => async (dispatch) => {
-  const jwtToken = useSelector((state) => state.userInfo.jwtToken);
+  const userInfo = await AsyncStorage.getItem('userInfo');
+  const { jwtToken } = JSON.parse(userInfo);
+
+  if (!jwtToken) {
+    throw new Error('JWT token not found. User must be logged in.');
+  }
+
   try {
     const response = await axios.get(`${URL}/users/favorites?page=0`, {
       headers: {
         Authorization: `Bearer ${jwtToken}`,
       },
     });
-
     if (response.status === 200) {
       const favoriteCars = response.data;
       dispatch(getFavoriteCars(favoriteCars));
@@ -252,10 +296,13 @@ export const fetchFavoriteCars = () => async (dispatch) => {
 
 export const fetchFilteredCars = (location, filters) => {};
 
-export const fetchPayments = (username) => {};
-
 export const fetchRentHistory = () => async (dispatch) => {
-  const jwtToken = useSelector((state) => state.userInfo.jwtToken);
+  const userInfo = await AsyncStorage.getItem('userInfo');
+  const { jwtToken } = JSON.parse(userInfo);
+
+  if (!jwtToken) {
+    throw new Error('JWT token not found. User must be logged in.');
+  }
   try {
     const response = await axios.get(`${URL}/users/details/bookings?page=0`, {
       headers: {
@@ -265,11 +312,49 @@ export const fetchRentHistory = () => async (dispatch) => {
     if (response.status === 200) {
       const rentHistory = response.data;
       dispatch(getRentHistory(rentHistory));
+      dispatch(fetchRentHistoryCars(rentHistory));
     } else {
       throw new Error('Fetching rent history failed.');
     }
   } catch (error) {
     console.error('Error during fetching rent history:', error);
+    throw error;
+  }
+};
+
+export const fetchRentHistoryCars = (rentHistory) => async (dispatch) => {
+  const userInfo = await AsyncStorage.getItem('userInfo');
+  const { jwtToken } = JSON.parse(userInfo);
+
+  if (!jwtToken) {
+    throw new Error('JWT token not found. User must be logged in.');
+  }
+
+  try {
+    const carsDetails = [];
+    for (const historyItem of rentHistory) {
+      const carId = historyItem['carId'];
+      try {
+        const response = await axios.get(`${URL}/cars/${carId}`, {
+          headers: {
+            Authorization: `Bearer ${jwtToken}`,
+          },
+        });
+
+        if (response.status === 200) {
+          const carDetails = response.data;
+          carsDetails.push(carDetails);
+        } else {
+          throw new Error(`Fetching details for carId ${carId} failed.`);
+        }
+      } catch (error) {
+        console.error(`Error during fetching details for carId ${carId}:`, error);
+        throw error;
+      }
+    }
+    dispatch(getRentHistoryCars(carsDetails));
+  } catch (error) {
+    console.error('Error during fetching rent history cars:', error);
     throw error;
   }
 };
